% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contains_multibyte_string.R
\name{contains_multibyte_string}
\alias{contains_multibyte_string}
\title{Tests whether a string contains a multibyte character}
\usage{
contains_multibyte_string(string, max_string_length = 256)
}
\arguments{
\item{string}{A character vector of length 1.}
}
\value{
TRUE/FALSE depending on whether the string contains a multibyte
}
\description{
Indicates whether a string contains a nasty multibyte character that throws an error in functions like tolower()
}
\author{
Ben Goehring \email{bengoehr@umich.edu}
}
