% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/um_colors.R
\docType{data}
\name{um_colors}
\alias{um_colors}
\title{um_colors}
\format{
A character vector of length 15.
}
\usage{
um_colors
}
\description{
A vector containing the University of Michigan's official colors
}
\examples{
um_colors
}
\keyword{datasets}
