% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uva_colors.R
\docType{data}
\name{uva_colors}
\alias{uva_colors}
\title{uva_colors}
\format{
A character vector of length 14.
}
\usage{
uva_colors
}
\description{
A vector containing the University of Virginia's official colors.
}
\examples{
uva_colors
}
\keyword{datasets}
